общий класс для viewModel фрагментов викторины
