+++ app/src/main/java/com/bartex/quizday/ui/flags/tabs/regions/IPreferenceHelper.kt
@@ -3,6 +3,4 @@ package com.bartex.quizday.ui.flags.tabs.regions
 interface IPreferenceHelper {
     fun savePositionState(position:Int)
     fun getPositionState(): Int
-    fun saveCurrentRegion(currentRegion:String)
-    fun getCurrentRegion():String
 }
