+++ app/src/main/java/com/bartex/quizday/ui/flags/tabs/regions/PreferenceHelper.kt
@@ -3,7 +3,6 @@ package com.bartex.quizday.ui.flags.tabs.regions
 import androidx.preference.PreferenceManager
 import com.bartex.quizday.App
 import com.bartex.quizday.model.common.Constants
-import com.bartex.quizday.ui.flags.tabs.regions.IPreferenceHelper
 
 class PreferenceHelper(val app: App): IPreferenceHelper {
 
@@ -21,16 +20,4 @@ class PreferenceHelper(val app: App): IPreferenceHelper {
         return position
     }
 
-    override fun saveCurrentRegion(currentRegion: String) {
-        PreferenceManager.getDefaultSharedPreferences(app)
-                .edit()
-                .putString(Constants.CURRENT_REGION, currentRegion)
-                .apply()
-    }
-
-    override fun getCurrentRegion():String{
-      val currentRegion =  PreferenceManager.getDefaultSharedPreferences(app)
-                .getString(Constants.CURRENT_REGION, Constants.REGION_ALL)!!
-        return currentRegion
-    }
 }
